requirements.txt
streamlit
pymongo
pandas
python-dotenv
langchain
langchain-groq
groq
bson
